v1.1.0/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0 